Correct a bug in AbsoluteLocationPath
According to the specification [1], a absolute location path expression
can consist of only a single slash (/). This is perhaps the most trivial
expression and is supposed to return the document root. 559d272
simplified the parser more than it should have.

This corrects the issue and perhaps illustrates a weakness in the suite.

[1] https://www.w3.org/TR/xpath/#NT-AbsoluteLocationPath